requirements.txt
fastapi==0.116.1
jinja2==3.1.6
Pillow==11.3.0
python-multipart==0.0.20
timm>=1.0.16,<2.0.0
torch==2.10.0
torchvision==0.25.0
transformers==4.56.1
uvicorn==0.35.0
